webapplication/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.3
django-bootstrap-form==3.4
django-filter==2.4.0
filelock==3.0.12
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
virtualenv==20.2.1
virtualenvwrapper-win==1.2.6
